dayflow-hrms/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
PyMySQL==1.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow>=10.0.0
cryptography>=3.4.8
mysql-connector-python>=8.0.0